// src/objects/enums.ts
/**
 * The available CornerPositions enum.
 */
export enum CornerPositions {
	none,
	topLeft,
	topRight,
	bottomLeft,
	bottomRight,
}

export enum BorderTypes {
	none = 'none',
	single = 'single',
	double = 'double',
	round = 'round',
	bold = 'bold',
	singleDouble = 'singleDouble',
	doubleSingle = 'doubleSingle',
	boldSingle = 'boldSingle',
	singleBold = 'singleBold',
}